+++ tests/integration/cases/call_deep_dotted_transformed/Tcl_call@v8_6.tcl
@@ -1,5 +1,0 @@
-proc app.client.fetch {args} {return {}}
-proc emit {args} {}
-emit [app.client.fetch "hello"]
-emit [app.client.fetch 42]
-emit [app.client.fetch 1]