+++ tests/integration/cases/call_dotted_transform_stub/Bash_call@v5_1.sh
@@ -1,5 +1,0 @@
-process() { :; }
-tracer.emit() { :; }
-tracer.emit "$(process "hello")"
-tracer.emit "$(process 42)"
-tracer.emit "$(process true)"