+++ tests/integration/cases/call_dotted_transform_stub/Php_call@v8_1.php
@@ -1,0 +1,7 @@
+<?php
+function process($value) {}
+class TracerType { function emit($_arg) {} }
+$tracer = new TracerType();
+tracer.emit(process(value: "hello"));
+tracer.emit(process(value: 42));
+tracer.emit(process(value: true));