+++ tests/integration/cases/call_dotted_transform_stub/PowerShell_call@v7.ps1
@@ -1,0 +1,6 @@
+function process {}
+class TracerType_ { [object] emit([object] $_arg) { return $null } }
+$tracer = [TracerType_]::new()
+tracer.emit(process("hello"))
+tracer.emit(process(42))
+tracer.emit(process($true))