+++ tests/integration/cases/call_keyword_args/Scheme_call@r7rs.scm
@@ -1,5 +1,0 @@
-(define throttler (lambda args 0))
-(define throttler.check (lambda args 0))
-(define emit (lambda args (if #f #f)))
-(emit (throttler.check "user_1" 1000.0))
-(emit (throttler.check "user_2" 2000.5))