+++ tests/integration/cases/call_no_params_transform/Clojure_call@v1_11.clj
@@ -1,4 +1,0 @@
-(defn process [& _args] 0)
-(defn emit [& _args] nil)
-(emit (process))
-(emit (process))