+++ tests/integration/cases/call_no_params_transform/Scheme_call@r7rs.scm
@@ -1,4 +1,0 @@
-(define process (lambda args 0))
-(define emit (lambda args (if #f #f)))
-(emit (process))
-(emit (process))