+++ tests/integration/cases/call_zip_values/JavaScript_call@es2015.js
@@ -1,0 +1,4 @@
+function process() {}
+function emit() {}
+emit(process({ value: "hello" }), true);
+emit(process({ value: 42 }), false);