+++ tests/integration/cases/call_zip_values/OCaml_call@v5.ml
@@ -1,0 +1,12 @@
+module Check = struct
+
+type val_t =
+  | OInt of int
+  | OStr of string
+  | OList of val_t list
+let process _ = ()
+let emit _ = ()
+let _ = emit(process("hello"), OBool true)
+let _ = emit(process(42), OBool false)
+
+end