+++ tests/integration/cases/call_zip_values/Rust_call@edition_2021.rs
@@ -1,0 +1,6 @@
+fn main() {
+    fn process<A>(_value: A) {}
+    fn emit<A, B>(__call: A, __zip: B) {}
+    emit(process("hello"), true);
+    emit(process(42), false);
+}